lib/indicators/indicators.hpp: fixing compile issue that appears all warnings on in ubuntu docker in macos

+++ lib/indicators/indicators.hpp
@@ -1424,13 +1424,14 @@ static inline int bisearch(wchar_t ucs, const struct interval *table, int max) {
   int min = 0;
   int mid;
 
-  if (ucs < table[0].first || ucs > table[max].last)
+  if (static_cast<int>(ucs) < table[0].first ||
+      static_cast<int>(ucs) > table[max].last)
     return 0;
   while (max >= min) {
     mid = (min + max) / 2;
-    if (ucs > table[mid].last)
+    if (static_cast<int>(ucs) > table[mid].last)
       min = mid + 1;
-    else if (ucs < table[mid].first)
+    else if (static_cast<int>(ucs) < table[mid].first)
       max = mid - 1;
     else
       return 1;
